fix: perform attest before cleanup (#35)
The cleanup step was being performed before the attest step, which
may have been causing the container digest to not be passed to the
attest step. This change moves the cleanup step to after the attest
step. It also updates the dev/container versions or whitespace to
trigger a build.

Fixes: #34
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>