feat: import release-trigger package
This commit imports the release-trigger package from the @jhatler
fork of Google's repo-automation-bots repository. It adds the package
as-is to the monorepo. It also adds a GitHub Actions workflow to
CI the package and integrates it with the monorepo's release process.

Refs: #48
URL: https://github.com/jhatler/repo-automation-bots
Commit: f3be771a2fe580ef2d7539989d21731810b2f1cc
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>