chore: release-trigger lint issues blocking import
This fixes the lint issues that are blocking the import of the
release-trigger package into the main project. This includes
those from the package itself and those from super-linter. ts-standard
was disabled in the super-linter configuration to allow the import
to go through (and it plays poorly with monorepos in general).

Refs: #48
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>