feat: import release-trigger package (#50)
* feat: import release-trigger package

This commit imports the release-trigger package from the @jhatler
fork of Google's repo-automation-bots repository. It adds the package
as-is to the monorepo. It also adds a GitHub Actions workflow to
CI the package and integrates it with the monorepo's release process.

Refs: #48
URL: https://github.com/jhatler/repo-automation-bots
Commit: f3be771a2fe580ef2d7539989d21731810b2f1cc
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>

* chore: release-trigger lint issues blocking import

This fixes the lint issues that are blocking the import of the
release-trigger package into the main project. This includes
those from the package itself and those from super-linter. ts-standard
was disabled in the super-linter configuration to allow the import
to go through (and it plays poorly with monorepos in general).

Refs: #48
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>

---------

Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>