feat: update release-trigger deps for node 22
This updates package.json and package-lock.json for the release-trigger
package to use the latest compatible versions of its dependencies. This
is necessary to support Node 22.

The dependencies updated are:

- @google-automations/bot-config-utils to ^7.0.0
- @google-automations/datastore-lock to ^5.0.0
- @google-automations/issue-utils to ^3.0.0
- @octokit/core to ^4.2.4
- @types/node to ^20.12.12
- gcf-utils to ^15.0.1
- jsonwebtoken to ^9.0.
- smee-client to ^1.2.5
- typescript to ^5.1.

Some changes to the release-trigger bot source and tests were needed to
support the split of `addOrUpdateIssueComment` from `gcf-utils` to
`issue-utils`.

Refs: #130
Signed-off-by: GitHub <noreply@github.com>