feat: add packer to build runner AMI (#274)
This adds a packer/ directory to the repository that will house the
files related to building machine images using Packer. The first
machine image that has been added is a starting point for a GitHub
Actions runner. This image is based on Ubuntu 22.04 and doesn't receive
any configuration at this time. That will be added in subsequent
commits.

This commit also adds a GitHub Actions workflow that will build the
runner AMI using Packer. This workflow is triggered by updates to the
packer/ directory (PR and push events). It also will run every 2 weeks
to ensure that the runner AMI is kept up to date.

Refs: #273
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>