feat: Add ansible support for Packer runner AMIs
This adds an ansible/ directory to the monorepo which will house the
various ansible playbooks and roles that this project uses. The first
ansible use case added is to support the creation of the runner AMIs
via Packer.

The runner use case contains two roles currently, `os` and `aws_cli`.

The `os` role is responsible for installing the basic packages that
the runner needs and for enabling multiarch support. The `aws_cli`
role is responsible for installing the AWS CLI and any plugins that
the runner needs.

These roles are included in the `site.yml` playbook which is intended
to be used exclusively by humans. Packer will use the `packer.yml`
playbook which includes the `site.yml` playbook. This makes it easier
to support the dynamic ansible inventories created by Packer.

The aws_ec2 inventory plugin has been configured so that running
instances can be manually managed. This also makes it easier to develop
with Ansible as you can iterate against the same instance without having
to recreate it as Packer does. The groups created by this plugin are the
primary way that variables are passed to the playbooks.

Refs: #275
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>