feat: Spacelift AWS integrations are self-managed
This extends the Spacelift control stack to include the ability to
configure the AWS integration. This is done by adding a new `aws.tf'
file that contains the necessary configuration for the AWS integration.

All child stacks have been moved to `stacks.tf` and this is where the
per-stack AWS Integration configuration will be done.

Refs: #279
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>