fix: apply cloud integrations after stack create
This updates the dependencies in the Spacelift control stack to ensure
that the cloud integrations are not added to a child stack until after
the stack has been created. This is necessary because the name of the
stacks is being defined via a local varable and not a resource
attribute.

Refs: #286
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>