feat: add network stack
This adds a simple network stack to Spacelift. It is intended to be
shared by other stacks that need a VPC, subnets, etc. The VPC contains
10 subnets across 3 availability zones. The subnets are divided into
public, DMZ, and internal subnets. The public subnets are intended for
load balancers, NAT, etc. Applications should be placed in the DMZ,
and data stores in the internal subnets.

Refs: #290
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>