feat: add stack for receiving webhooks (#337)
This adds a `webhooks` stack that will be used to receive webhooks
related to the control repository. New roles and policies are added to
the `auth` stack to allow the `webhooks` stack to interact with SQS,
CloudWatch, DynamoDB, and EC2.

The initial implementation of the `webhooks` stack simply receives the
payload from the webhook and writes it to SQS. The `webhooks` stack will
have a Lambda function that will be triggered by the SQS message and
will process the webhook payload. This needs added on a subsequent PR.

Fixes: #336
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>