feat: actions agent setup by ansible on runners
This commit adds the ansible roles and playbooks to setup the GitHub
Actions runner agent via Packer or AWS SSM. It also normalizes the
naming around the runners so that the only singular usage of "runner"
apply to specific instances such as AMIs or Launch Templates. The
resources which will apply to multiple runners are now named "runners"
(the plural form).

Additionally, this commit adds the resources to the `runners` Spacelift
stack to define launch templates for the runners. These templates are
defined using the runner-template module from this repo.

Refs: #362
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>