feat: add ubuntu cloud image support
This adds a workflow to build standardized Ubuntu Cloud Images. AMIs are
build and tested using Packer. They are pre-configured with cloud-init
and the AWS SSM agent. The images are built using the official Ubuntu
Cloud Image artifacts.

Containers are built using buildah The containers undergo normalization
and minimalization to ensure they are as small as possible and behave
consistently.

The ubuntu-cloud-images script is used to scope the workflow to only
the Ubuntu Cloud Image release which it has not yet built. This script
caches the etag of each release so that it can be skipped if it has
already been built. The script outputs this scope to json as a step
output for use as a matrix strategy in the workflow. It also securely
downloads the official Ubuntu Cloud Image artifacts.

Fixes: #378
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>
Signed-off-by: GitHub <noreply@github.com>