feat: container PR workflow on self-hosted runners
This migrates the container PR workflow to self-hosted runners. This
was done by introducing a new CI methodology whereby the various
container builds are dynamically generated for the workflow matrix. This
was done by introducing a new `.ci.py` script that is used to generate
the workflow matrix by walking the project tree for other `.ci.*` files
in the changed paths.

Fixes: #210
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>